agenticService/requirements.txt
fastapi
uvicorn
google-genai
langgraph
langchain-google-genai
python-dotenv
